+++ SSAS_Project/app/src/main/java/com/example/ssas_project/DeleteStudentActivity.java
@@ -6,6 +6,7 @@ import android.view.View;
 import android.widget.EditText;
 import android.widget.Button;
 import android.widget.ListView;
+import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -64,11 +65,22 @@ public class DeleteStudentActivity extends AppCompatActivity{
                 @Override
                 public void onClick(View view) {
                     String del_student = del_student_id1.getText().toString();
-                    myDAO.deleteEnroll(Integer.parseInt(offering_id1), Integer.parseInt(del_student));
-                    Intent intent = new Intent(DeleteStudentActivity.this, Class_Info_Activity.class);
-                    intent.putExtra("course_id", course_id1);
-                    intent.putExtra("offer_id", offering_id1);
-                    startActivity(intent);
+                    if(del_student.isEmpty()){
+                        Toast.makeText(DeleteStudentActivity.this, "Please enter the required field!", Toast.LENGTH_SHORT).show();
+                    }
+                    else {
+                        List<Integer> check_list = myDAO.getEnrollStudents(Integer.parseInt(offering_id1));
+                        if(check_list.contains(Integer.parseInt(del_student))){
+                            myDAO.deleteEnroll(Integer.parseInt(offering_id1), Integer.parseInt(del_student));
+                            Intent intent = new Intent(DeleteStudentActivity.this, Class_Info_Activity.class);
+                            intent.putExtra("course_id", course_id1);
+                            intent.putExtra("offer_id", offering_id1);
+                            startActivity(intent);
+                        }
+                        else{
+                            Toast.makeText(DeleteStudentActivity.this, "Student not in the class!", Toast.LENGTH_SHORT).show();
+                        }
+                    }
                 }
             });
